+++ changelog.md
@@ -17,6 +17,7 @@
 - Fixed an annoying consolespam with the Cartographer trades (BluSunrize)
 - Fixed various GUIs still using "RF" instead of "IF" (BluSunrize)
 - Fixed Lightningrod resulting in a crash on very specific setups (BluSunrize)
+- Fixed dupebug with ArcFurnace recycling (BluSunrize)
 - Translations Added/Updated: ja_jp (karakufire), es_es(Dorzar & rogama25)
 
 #####Version 0.12-86 - BUILT